Create selected ability processor class
선택된 능력 처리 담당 클래스 생성.
해당 클래스는 선택된 능력의 인덱스 판단 후, 플레이어에게 그에 맞는 효과 부여 및 수치 조정을 수행.
